% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-archiver.R
\name{archive_data}
\alias{archive_data}
\title{Create or maintain a data archive structure}
\usage{
archive_data(path, season, clean_build = FALSE, verbose = TRUE)
}
\arguments{
\item{path}{Where to generate the data archive}

\item{season}{Which season to generate data for (e.g., 20242025)}

\item{clean_build}{Whether to create or overwrite existing data, defaults to FALSE}

\item{verbose}{Get chatty? Defaults to TRUE}
}
\description{
This function will generate a set of season-level data, in CSV and JSON format
This will include things like teams, franchises, boxscores, play-by-play data, and rosters
}
