% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-draft-results.R
\name{build_draft_results}
\alias{build_draft_results}
\title{Build draft_results table}
\usage{
build_draft_results(conn = connect_to_db())
}
\arguments{
\item{conn}{A connection to a database - defaults to creating a new one}
}
\description{
The internals here are hard-coded, as this is not meant to be
something used outside of the current process, at the moment.
In the future, perhaps something config-based could be created
This is meant to create the table brand new, not update it. If the table already
exists, it will be re-written by running this function.
}
