% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-new-nhl-db.R
\name{build_new_nhl_db}
\alias{build_new_nhl_db}
\title{Function to build NHL database from scratch}
\usage{
build_new_nhl_db(seasons, dbname = "hockey", schema = "nhl",
  num_cores = 1, conn = connect_to_db())
}
\arguments{
\item{seasons}{Which seasons to}

\item{dbname}{A Database name to put the tables.}

\item{schema}{A schema within that database to put the tables.}

\item{num_cores}{The number of cores, if choosing to parallelize (default = 1)}

\item{conn}{String, takes value 'sqlite' or -- a default one is generated if not provided.  Defaults to 'postgres'}
}
\description{
This is a wrapper for the ordered running of a specific set of functions
aimed at generating a brand new database.
However!  That database and schema need to _already_ exist, or this will fail (or will it?).
}
