% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-player-info-table.R
\name{build_player_info_table}
\alias{build_player_info_table}
\title{Build player_info table}
\usage{
build_player_info_table(playerIDs, conn = connect_to_db())
}
\arguments{
\item{playerIDs}{a vector of playerIDs to process and add to the database}

\item{conn}{A connection to a database - defaults to creating a new one}
}
\description{
The internals here are hard-coded, as this is not meant to be
something used outside of the current process, at the moment.
In the future, perhaps something config-based could be created
This is meant to create the table brand new, not update it. If the table already
exists, it will be re-written by running this function.
}
