% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-schedule-table.R
\name{build_schedule_table}
\alias{build_schedule_table}
\title{Build schedule table}
\usage{
build_schedule_table(seasons, conn = connect_to_db(), return_vals = FALSE)
}
\arguments{
\item{seasons}{A vector of seasons (format: 20182019)}

\item{conn}{A connection to a database - defaults to creating a new one}

\item{return_vals}{Logical. If true, the tables are returned}
}
\description{
The internals here are hard-coded, as this is not meant to be
something used outside of the current process, at the moment.
In the future, perhaps something config-based could be created
This is meant to create the table brand new, not update it. If the table already
exists, it will be re-written by running this function.
}
