% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-play-by-play.R
\name{clean_pbp_data}
\alias{clean_pbp_data}
\title{Clean pbp data}
\usage{
clean_pbp_data(d)
}
\arguments{
\item{d}{pbp_data}
}
\description{
Clean pbp data
}
\details{
Cleans up some column names, converts a few things
}
