% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning-utils.R
\name{clean_player_data}
\alias{clean_player_data}
\title{Clean player data}
\usage{
clean_player_data(players)
}
\arguments{
\item{players}{Data.table containing skater info}
}
\description{
The purpose of this function is to clean player data.  This includes removing unwanted columns,
renaming wanted columns, and replacing NAs with 0.
}
