% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning-utils.R
\name{clean_team_stat_data}
\alias{clean_team_stat_data}
\title{Clean team stat data}
\usage{
clean_team_stat_data(team_stats)
}
\arguments{
\item{team_stats}{Data.table containing team stats at the game level}
}
\description{
Clean team stat data
}
