% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect-to-db.R
\name{connect_to_db}
\alias{connect_to_db}
\title{Connect to DB}
\usage{
connect_to_db(
  type = "sqlite",
  db_path = "tempdb.sqlite",
  dbname = "hockey",
  schema = "nhl",
  host = "localhost",
  port = 5432,
  user = "postgres",
  password = getPass::getPass()
)
}
\arguments{
\item{type}{One of "sqlite" or "postgres" (defaults to "sqlite" as an internal test)}

\item{db_path}{Location of the SQLite DB -- ignored if type is 'postgres'}

\item{dbname}{The database name. If type = 'sqlite', it will be created. If type = 'postgres', it must be an existing database. Defaults to "hockey"}

\item{schema}{Only used if type = 'postgres'. Defaults to 'nhl'}

\item{host}{Defaults to "localhost"}

\item{port}{Defaults to 5432}

\item{user}{Defaults to 'postgres'}

\item{password}{Defaults to getPass::getPass()}
}
\value{
an active connection to a db called "tempdb"
}
\description{
Currently a shell for a SQLite connection
}
