% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning-utils.R
\name{convert_ftinch}
\alias{convert_ftinch}
\title{Convert feet/inches to inches}
\usage{
convert_ftinch(ftinch)
}
\arguments{
\item{ftinch}{string of format `ft' in"`}
}
\description{
Convert feet/inches to inches
}
