% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning-utils.R
\name{convert_time}
\alias{convert_time}
\title{Convert time}
\usage{
convert_time(time_col)
}
\arguments{
\item{time_col}{Column containing "MM:SS" formatted data, as is typically provided by NHL api}
}
\description{
Convert time
}
