% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-db.R
\name{create_db_from_archive}
\alias{create_db_from_archive}
\title{Create Database from Archive}
\usage{
create_db_from_archive(
  archive_path,
  database_connection = NULL,
  local_path = "."
)
}
\arguments{
\item{archive_path}{Path to existing data archive}

\item{database_connection}{Connection to Database}

\item{local_path}{If no database connection is passed, we'll build a local one in 'csv' format in this path}
}
\description{
This function uses a data archive, built using `archive_data` function
calls and NHL API calls, to generate a relational database.  That means
the database will contain as many years as are present in your data
archive path. It also means the database building process involves a clean build every
time, not incremental. So, to update the database, you first update the
archive, then run this script again.
}
