% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-player-info.R
\name{fetch_all_players}
\alias{fetch_all_players}
\title{Function to fetch player info for ALL players}
\usage{
fetch_all_players()
}
\value{
Data.table containing all players of all time
}
\description{
This function takes no arguments. It provides a method of grabbing basic data about every player in the NHL database.
}
