% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-team-data.R
\name{fetch_all_teams}
\alias{fetch_all_teams}
\title{Function to get List of Teams}
\usage{
fetch_all_teams(nhl_only = TRUE, active_only = TRUE)
}
\arguments{
\item{nhl_only}{Defaults to TRUE, wherein the function returns only NHL teams
If FALSE, you'll teams in other leagues (AHL, OHL, etc.)}

\item{active_only}{Defaults to TRUE, wherein you'll only get active teams.}
}
\description{
Function to get List of Teams
}
