% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-boxscore.R
\name{fetch_boxscores}
\alias{fetch_boxscores}
\title{Function to fetch game boxcores}
\usage{
fetch_boxscores(gameID, verbose = FALSE)
}
\arguments{
\item{gameID}{A vector of gameIDs}

\item{verbose}{Default FALSE}
}
\description{
Function to fetch game boxcores
}
\details{
Returns the raw list-formatted JSON of boxscore data
}
