% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-draft.R
\name{fetch_draft_data}
\alias{fetch_draft_data}
\title{Fetch the draft data}
\usage{
fetch_draft_data(year = NULL)
}
\arguments{
\item{year}{Draft Year - If not provided, it will just return every draft ever}
}
\description{
Fetch the draft data
}
\details{
Automatically grabs all historical draft data back to the 1960s.
}
