% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-game-data.R
\name{fetch_game_data}
\alias{fetch_game_data}
\title{Function to fetch game data}
\usage{
fetch_game_data(gameID, verbose = FALSE)
}
\arguments{
\item{gameID}{A vector of gameIDs}

\item{verbose}{Default FALSE}
}
\description{
Function to fetch game data
}
\details{
Returns game data. Involves hitting two endpoints to get high level scoring, penalty, and team-stat data
}
