% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-schedule.R
\name{fetch_game_ids}
\alias{fetch_game_ids}
\title{Function to get game IDs for a given time period}
\usage{
fetch_game_ids(season = NULL, season_type = 2)
}
\arguments{
\item{season}{Season to query (e.g., '20152016'). Passing nothing will return the current day}

\item{season_type}{The season type, where 1 = preseason, 2 = regular season, 3 = playoffs, 4 = all-star (Default 2)}
}
\description{
Function to get game IDs for a given time period
}
\details{
If season, start_date, and end_date are NOT provided, the default will be the current date
}
