% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-player-info.R
\name{fetch_player_info}
\alias{fetch_player_info}
\title{Function to fetch player info for a single player}
\usage{
fetch_player_info(playerID)
}
\arguments{
\item{playerID}{A player ID to fetch}
}
\description{
Function to fetch player info for a single player
}
