% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-prospect-info.R
\name{fetch_prospect_info}
\alias{fetch_prospect_info}
\title{Function to fetch prospect data}
\usage{
fetch_prospect_info(verbose = FALSE)
}
\arguments{
\item{verbose}{whether to provide messaging, default FALSE}
}
\description{
Function to fetch prospect data
}
\details{
only provides most recent crop of prospects, not historical data
}
