% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-team-data.R
\name{fetch_roster_data}
\alias{fetch_roster_data}
\title{Function to fetch Roster Data}
\usage{
fetch_roster_data(season, team)
}
\arguments{
\item{season}{Season (e.g., 20242025)}

\item{team}{Team Abbreviation (e.g., PIT)}
}
\description{
Function to fetch Roster Data
}
\details{
Simply provides rosters for a given season
}
