% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-shifts.R
\name{fetch_shift_data}
\alias{fetch_shift_data}
\title{Fetch Shift level information}
\usage{
fetch_shift_data(gameID, verbose = FALSE)
}
\arguments{
\item{gameID}{Game ID for shift data}

\item{verbose}{Defaults to FALSE}
}
\description{
Fetch Shift level information
}
