% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-team-data.R
\name{fetch_team_data}
\alias{fetch_team_data}
\title{Function to get all teams data, including rosters!}
\usage{
fetch_team_data(season = NULL)
}
\arguments{
\item{season}{Season to pull roster data for, accepts one (format '20182019')}
}
\description{
Function to get all teams data, including rosters!
}
