% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-time-on-ice.R
\name{fetch_time_on_ice}
\alias{fetch_time_on_ice}
\title{Function to fetch time on ice report}
\usage{
fetch_time_on_ice(gameID, verbose = FALSE)
}
\arguments{
\item{gameID}{A vector of gameIDs}

\item{verbose}{Default FALSE}
}
\description{
Function to fetch time on ice report
}
\details{
Returns the raw list-formatted JSON of TOI data
}
