% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-fantasy-points.R
\name{generate_fantasy_points}
\alias{generate_fantasy_points}
\title{Generate fantasy points}
\usage{
generate_fantasy_points(players, site = "draftkings", type = NULL)
}
\arguments{
\item{players}{List of players with stats.}

\item{site}{Currently only "draftkings" is supported}

\item{type}{Either 'goalie' or 'skater'; if not provided, we'll try to guess.}
}
\description{
Generate fantasy points
}
