% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goalies.R
\docType{data}
\name{goalies}
\alias{goalies}
\title{NHL Goalie Statistics 2007-2019}
\format{A data frame with 856 rows and 16 variables:
\describe{
  \item{player}{Player name}
  \item{season}{Season}
  \item{team}{First team player played for in a given season}
  \item{team2}{Second team player played for in a given season}
  \item{team3}{Third team player played for in a given season}
  \item{gp}{Games Played}
  \item{toi}{Time on the ice}
  \item{sa}{Shots against}
  \item{ga}{Goals against}
  \item{sv_pct}{Save percentage}
  \item{xsv_pct}{Expected save percentage}
  \item{dsv_pct}{Dangerous save percentage}
  \item{ldsv_pct}{Low danger save percentage}
  \item{mdsv_pct}{Medium danger save percentage}
  \item{hdsv_pct}{High danger save percentage}
  \item{gsaa}{Goals saved above average}
}}
\source{
\url{http://corsica.hockey/goalie-stats/}
}
\usage{
data(goalies)
}
\description{
A dataset containing season-level statistics for NHL goalies for all
game states (5v5, PP, PK). Includes regular season and playoffs.
}
\keyword{datasets}
