% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-boxscore.R
\name{parse_boxscore_json}
\alias{parse_boxscore_json}
\title{Parse the basic boxscore data}
\usage{
parse_boxscore_json(box_file)
}
\arguments{
\item{box_file}{JSON file containing boxscore data}
}
\description{
Parse the basic boxscore data
}
