% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-draft.R
\name{parse_draft_data}
\alias{parse_draft_data}
\title{Convert Draft data into a table}
\usage{
parse_draft_data(dd)
}
\arguments{
\item{dd}{draft data from `fetch_draft_data()`}
}
\description{
Convert Draft data into a table
}
