% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-game-data.R
\name{parse_game_json}
\alias{parse_game_json}
\title{Parse Game Data JSON}
\usage{
parse_game_json(game_json)
}
\arguments{
\item{game_json}{Path to a JSON file containing the Game-Summary data}
}
\description{
Parse Game Data JSON
}
