% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-play-by-play.R
\name{parse_pbp}
\alias{parse_pbp}
\title{Function to parse play-by-play data}
\usage{
parse_pbp(pbp)
}
\arguments{
\item{pbp}{Play by play data from NHL api}
}
\description{
Function to parse play-by-play data
}
