% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-schedule.R
\name{parse_schedule}
\alias{parse_schedule}
\title{Function to parse schedule data}
\usage{
parse_schedule(schedule_json)
}
\arguments{
\item{schedule_json}{The JSON output from NHL Schedule API}
}
\description{
Function to parse schedule data
}
