% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player-stats-boxscore-DEP.R
\name{player_stats_from_boxscore}
\alias{player_stats_from_boxscore}
\title{function to parse game boxscore data for player stats}
\usage{
player_stats_from_boxscore(boxscore, gid)
}
\arguments{
\item{boxscore}{Parsed content from an NHL api call to boxscore for a single game}

\item{gid}{GameId -- comes along for safekeeping!}
}
\description{
function to parse game boxscore data for player stats
}
