% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-archive.R
\name{process_archive}
\alias{process_archive}
\title{Process Archive ()}
\usage{
process_archive(archive_path, process_path = NULL)
}
\arguments{
\item{archive_path}{Path to existing data archive}

\item{process_path}{Where to save the processed data (defaults to subfolder in archive_path)}
}
\description{
This function uses a data archive, built using `archive_data` function
calls and NHL API calls, to create a set of files (current in CSV) that
can easily be converted into a relational database of your choosing. It is run
at the SEASON level (so, archive/20242025/ as an example)
}
