% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team-stats-boxscore-DEP.R
\name{team_stats_from_boxscore}
\alias{team_stats_from_boxscore}
\title{function to parse game boxscore data for team stats}
\usage{
team_stats_from_boxscore(boxscore)
}
\arguments{
\item{boxscore}{Parsed content from an NHL api call to boxscore for a single game}

\item{gid}{the game ID, used to keep track of what's going on.}
}
\description{
function to parse game boxscore data for team stats
}
