% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\docType{data}
\name{teams}
\alias{teams}
\title{NHL Team Statistics 2007-2019}
\format{A data frame with 362 rows and 28 variables:
\describe{
  \item{team}{Team name}
  \item{season}{Season}
  \item{gp}{Games Played}
  \item{toi}{Time on the ice}
  \item{cf}{Corsi for}
  \item{ca}{Corsi against}
  \item{c_plumin}{Corsi plus/minus (cf-ca)}
  \item{cf_60}{Corsi for per 60 minutes on the ice}
  \item{ca_60}{Corsi against per 60 minutes on the ice}
  \item{gf}{Goals scored for a player's team}
  \item{ga}{Goals scored against player's team}
  \item{g_plumin}{Goals for - goals against}
  \item{gf_pct}{Percentage of all goals scored by team}
  \item{gf_60}{Goals for per 60 minutes on the ice}
  \item{ga_60}{Goals against per 60 minutes on the ice}
  \item{xgf}{Expected goals for}
  \item{xga}{Expected goals against}
  \item{xg_plumin}{Expected goals for - expected goals against}
  \item{xgf_pct}{Expected goals for as a percentage of total expected goals for}
  \item{xgf_60}{Expected goals for per 60 minutes on the ice}
  \item{xga_60}{Expected goals against per 60 minutes on the ice}
  \item{pent}{Penalties taken}
  \item{pend}{Penalties drawn}
  \item{p_plumin}{Penalties taken - indiviual penalties drawn}
  \item{sh_pct}{Shooting percentage}
  \item{sv_pct}{Save percentage}
  \item{pdo}{Just Win Baby}
}}
\source{
\url{http://corsica.hockey/team-stats/}
}
\usage{
data(teams)
}
\description{
A dataset containing season-level statistics for NHL teams for all
game states (5v5, PP, PK). Includes regular season.
}
\keyword{datasets}
