% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-player-info-table.R
\name{update_player_info_table}
\alias{update_player_info_table}
\title{Update the player_info table}
\usage{
update_player_info_table(gameIDs = NULL, conn)
}
\arguments{
\item{gameIDs}{Game IDs from which to grab playerids (Optional, used to limit scope)}

\item{conn}{A db connection. There is no default.}
}
\description{
only necessary due to possible additions to the table.
Relies on calling existing tables.
}
