% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-plays-tables.R
\name{update_plays_tables}
\alias{update_plays_tables}
\title{Update plays tables}
\usage{
update_plays_tables(dat = NULL, conn, gameIDs = NULL)
}
\arguments{
\item{dat}{A data set for updating. Can be NULL, otherwise, should be the output of lapply(gameIDs, fetch_boxscore_stats)}

\item{conn}{a DB connection (required)}

\item{gameIDs}{if dat is null, the function will try to use "gameIDs" -- a vector of game ids!}
}
\description{
One of either dat or gameIDs MUST be provided.
}
