% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-schedule-table.R
\name{update_schedule_table}
\alias{update_schedule_table}
\title{Update schedule table}
\usage{
update_schedule_table(dat = NULL, conn, ...)
}
\arguments{
\item{dat}{A data set for updating. Can be NULL.}

\item{conn}{a DB connection (required)}

\item{...}{additional arguments passed to 'build_schedule_table'}
}
\description{
Update schedule table
}
